refactor: component files to better visualization
