+++ Front/src/Pages/ManagePage/DeleteConfirmationModal.jsx
@@ -1,0 +1,40 @@
+import { motion } from 'framer-motion';
+
+const DeleteConfirmationModal = ({ showModal, onClose, onConfirm }) => {
+  if (!showModal) return null;
+
+  return (
+    <motion.div
+      className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center"
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      exit={{ opacity: 0 }}
+    >
+      <motion.div
+        className="bg-white p-6 rounded-lg shadow-lg w-full max-w-md"
+        initial={{ scale: 0.8 }}
+        animate={{ scale: 1 }}
+        transition={{ duration: 0.2 }}
+      >
+        <h2 className="text-xl font-bold mb-4">Confirmar Deleção</h2>
+        <p className="mb-6">Tem certeza que deseja deletar este item? Essa ação não pode ser desfeita.</p>
+        <div className="flex justify-end space-x-4">
+          <button
+            onClick={onClose}
+            className="px-4 py-2 text-sm font-medium text-gray-700 bg-gray-200 rounded-md hover:bg-gray-300"
+          >
+            Cancelar
+          </button>
+          <button
+            onClick={onConfirm}
+            className="px-4 py-2 text-sm font-medium text-white bg-red-600 rounded-md shadow-sm hover:bg-red-700"
+          >
+            Deletar
+          </button>
+        </div>
+      </motion.div>
+    </motion.div>
+  );
+};
+
+export default DeleteConfirmationModal;
