+++ Front/src/main.jsx
@@ -16,9 +16,9 @@ import SaibaMais from './Pages/Projetos/SaibaMais/SaibaMais.jsx';
 import Membros from './Pages/Equipes/Membros.jsx';
 import Login from './Pages/Login/Login.jsx';
 import Signup from './Pages/Signup/Signup.jsx';
-import CreateProject from './Pages/ManageProjects/CreateProject/CreateProject.jsx';
-import ManageProjects from './Pages/ManageProjects/ManageProjects.jsx';
-import EditProject from './Pages/ManageProjects/EditProject/EditProject.jsx';
+import CreateProject from './Pages/ManagePage/ManageProjects/CreateProject/CreateProject.jsx'
+import ManagePage from './Pages/ManagePage/ManagePage.jsx';
+import EditProject from './Pages/ManagePage/ManageProjects/EditProject/EditProject.jsx'
 
 const router = createBrowserRouter([
   {
@@ -60,7 +60,7 @@ const router = createBrowserRouter([
   },
   {
     path: "/manage",
-    element: <ManageProjects/>
+    element: <ManagePage/>
   },
   {
     path: "/manage/create-project",
